[Fact] enum 클래스 사용으로 총상금 더하기
